Fix debugService instance methods - remove static keyword
The debugService methods were defined as static but we export an instance.
Static methods belong to the class, not instances. Converted all methods to
instance methods so they can be called on the exported debugService instance.

Methods changed from static to instance:
- isDevEnvironment()
- toggleDebugPremium()
- isDebugPremiumEnabled()
- resetDebugSettings()
- getDebugInfo()

This fixes the error: "debugService.isDebugPremiumEnabled is not a function"